split simulation methods into separate files
