+++ lhotse/workflows/meeting_simulation/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseMeetingSimulator
+from .conversational import ConversationalMeetingSimulator
+from .speaker_independent import SpeakerIndependentMeetingSimulator
